requirements.txt
blinker==1.7.0
certifi==2024.8.30
charset-normalizer==2.0.12
click==8.1.7
config==0.5.1
Flask~=3.0.3
Flask-Cors==4.0.0
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
requests~=2.32.3
urllib3==1.26.20
Werkzeug==3.0.2
python-dotenv~=1.0.1